Safari-611.1.10/LayoutTests/svg/dom/SVGStyleElement/style-langspace.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description("This test checks that SVGStyleElement objects have xmlspace and xmllang properties inherited from SVGLangSpace.");

var style = document.createElementNS("http://www.w3.org/2000/svg", "style");

style.xmlspace = "preserve";
shouldBeEqualToString("style.xmlspace", "preserve");

style.xmllang = "en-AU";
shouldBeEqualToString("style.xmllang", "en-AU");

successfullyParsed = true;
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
