third_party/blink/web_tests/platform/generic/svg/dom/SVGStyleElement/style-langspace-expected.txt

This test checks that SVGStyleElement objects have xmlspace and xmllang properties inherited from SVGLangSpace.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS style.xmlspace is "preserve"
PASS style.xmllang is "en-AU"
PASS successfullyParsed is true

TEST COMPLETE
